Update the `StakeData` entity
Track the amount of legacy and native stake. Add mapping functions to
update amounts. NOTE: the `callHandlers` doesn't work with local Ganache
chain maybe we may be able to recognize which token has been unstaked or
topped-up from event data. Need to dig into it and figure out how to
test it locally.