Remove unnecessary `callHandlers` from manifest
The `callHandlers` only works with ethereum nodes that support the
`trace_filter` RPC method. The Ganache node doesn't support this method
and it not possible to test these handlers locally with local chain.

The `TTokenStaking` contract supports three types of top-ups(T/KEEP/NU)
and four types of unstakes(T/KEEP/NU/ALL) and each type is a separate
contract function but emits the same event w/o the type param. So to
recoginze which function has been called we extract first 4 bytes from
event input data, which are the function identifier, and compare with
the all types of topup and unstake function identifieres. If none of the
function identifiers match, we get the staking balance directly from the
chain and save in the subgrph entity.