Add utils function to populate new epoch stakes
To create a new Epoch, it's necessary to copy the last epoch stakes but
replacing the ID for a new one. This task is made in at least three
handlers (staked, toppedup and unstaked), so a new function has been
created in utils.

Also, a new constant variable for amount and staking provider has been
added to staked and unstaked handler functions.