Use `SimplePREApplication` name for schema
The contract used for PRE application is 'SimplePREApplication' instead
of 'PREApplication' since the current version of the app is not the
definitive one.

Also, this GraphQL entity includes a field for 'StakeData', and two
timestamps fields, one for when the operator is bonded and another for
when the operator is confirmed.