Fix operator bonded and confirmed timestamps
Up to now, the field `bondedTimestamp` and `confirmedTimestamp` on
`SimplePREApplication` subgraph entity contains the timestamps of the
last event received. This can be a problem if a staker changes its
operator since it will be taken this timestamp for rewards calculation,
so the time in which the previous operator was running will not be taken
into account.

The fix consists of not overwriting these timestamps when a new `bonded
operator` or `confirmed operator` event is received.